DTCCPT-claude-dtc-assessment-tool-6jbwq/requirements.txt
# DTC AI Agent Capability Assessment Tool
# Core dependencies

# Web Framework
streamlit>=1.31.0

# Anthropic Claude API
anthropic>=0.18.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0

# Data Processing
pyyaml>=6.0
pydantic>=2.0.0

# Export capabilities
markdown>=3.5.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
